+++ ViewUser.js
@@ -1,4 +1,4 @@
-import _ from 'lodash';
+import { cloneDeep, get, omit, differenceBy } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import queryString from 'query-string';
@@ -29,8 +29,6 @@ import {
   UserLoans,
 } from './lib/ViewSections';
 
-import css from './UserForm.css';
-
 class ViewUser extends React.Component {
 
   static propTypes = {
@@ -49,6 +47,9 @@ class ViewUser extends React.Component {
         records: PropTypes.arrayOf(PropTypes.object),
       }),
       patronGroups: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
+      settings: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
       }),
     }),
@@ -115,6 +116,11 @@ class ViewUser extends React.Component {
       path: 'perms/users/:{id}/permissions',
       params: { indexField: 'userId' },
     },
+    settings: {
+      type: 'okapi',
+      records: 'configs',
+      path: 'configurations/entries?query=(module=USERS and configName=profile_pictures)',
+    },
   });
 
   constructor(props) {
@@ -126,6 +132,7 @@ class ViewUser extends React.Component {
       selectedLoan: {},
       lastUpdate: null,
       sections: {
+        userInformationSection: true,
         extendedInfoSection: false,
         contactInfoSection: false,
         proxySection: false,
@@ -213,7 +220,7 @@ class ViewUser extends React.Component {
 
   // eslint-disable-next-line class-methods-use-this
   getUserFormData(user, addresses, sponsors, proxies, permissions) {
-    const userForData = user ? _.cloneDeep(user) : user;
+    const userForData = user ? cloneDeep(user) : user;
     userForData.personal.addresses = addresses;
     Object.assign(userForData, { sponsors, proxies, permissions });
 
@@ -231,7 +238,7 @@ class ViewUser extends React.Component {
     if (sponsors) this.props.updateSponsors(sponsors);
     if (permissions) this.updatePermissions(permissions);
 
-    const data = _.omit(user, ['creds', 'proxies', 'sponsors', 'permissions']);
+    const data = omit(user, ['creds', 'proxies', 'sponsors', 'permissions']);
 
     this.props.mutator.selUser.PUT(data).then(() => {
       this.setState({
@@ -244,8 +251,8 @@ class ViewUser extends React.Component {
   updatePermissions(perms) {
     const mutator = this.props.mutator.permissions;
     const prevPerms = (this.props.resources.permissions || {}).records || [];
-    const removedPerms = _.differenceBy(prevPerms, perms, 'id');
-    const addedPerms = _.differenceBy(perms, prevPerms, 'id');
+    const removedPerms = differenceBy(prevPerms, perms, 'id');
+    const addedPerms = differenceBy(perms, prevPerms, 'id');
     eachPromise(addedPerms, mutator.POST);
     eachPromise(removedPerms, mutator.DELETE);
   }
@@ -256,7 +263,7 @@ class ViewUser extends React.Component {
   // value. If so, this returns a locally stored update date until the data
   // is refreshed.
   dateLastUpdated(user) {
-    const updatedDateRec = _.get(user, ['updatedDate'], '');
+    const updatedDateRec = get(user, ['updatedDate'], '');
     const updatedDateLocal = this.state.lastUpdate;
 
     if (!updatedDateRec) { return ''; }
@@ -273,7 +280,7 @@ class ViewUser extends React.Component {
 
   handleSectionToggle({ id }) {
     this.setState((curState) => {
-      const newState = _.cloneDeep(curState);
+      const newState = cloneDeep(curState);
       newState.sections[id] = !newState.sections[id];
       return newState;
     });
@@ -281,7 +288,7 @@ class ViewUser extends React.Component {
 
   handleExpandAll(obj) {
     this.setState((curState) => {
-      const newState = _.cloneDeep(curState);
+      const newState = cloneDeep(curState);
       newState.sections = obj;
       return newState;
     });
@@ -293,6 +300,7 @@ class ViewUser extends React.Component {
     const user = this.getUser();
     const patronGroups = (resources.patronGroups || {}).records || [];
     const permissions = (resources.permissions || {}).records || [];
+    const settings = (resources.settings || {}).records || [];
     const sponsors = this.props.getSponsors();
     const proxies = this.props.getProxies();
 
@@ -309,16 +317,16 @@ class ViewUser extends React.Component {
       </Pane>
     );
 
-    const patronGroupId = _.get(user, ['patronGroup'], '');
+    const patronGroupId = get(user, ['patronGroup'], '');
     const patronGroup = patronGroups.find(g => g.id === patronGroupId) || { group: '' };
-    const addresses = toListAddresses(_.get(user, ['personal', 'addresses'], []), this.addressTypes);
+    const addresses = toListAddresses(get(user, ['personal', 'addresses'], []), this.addressTypes);
     const userFormData = this.getUserFormData(user, addresses, sponsors, proxies, permissions);
 
     return (
-      <Pane id="pane-userdetails" defaultWidth={this.props.paneWidth} paneTitle={<span><Icon icon="profile" iconRootClass={css.UserFormEditIcon} /> {getFullName(user)}</span>} lastMenu={detailMenu} dismissible onClose={this.props.onClose}>
+      <Pane id="pane-userdetails" defaultWidth={this.props.paneWidth} paneTitle={getFullName(user)} lastMenu={detailMenu} dismissible onClose={this.props.onClose}>
         <Row end="xs"><Col xs><ExpandAllButton accordionStatus={this.state.sections} onToggle={this.handleExpandAll} /></Col></Row>
 
-        <UserInfo stripes={stripes} user={user} patronGroup={patronGroup} />
+        <UserInfo stripes={stripes} user={user} patronGroup={patronGroup} accordionId="userInformationSection" expanded={this.state.sections.userInformationSection} onToggle={this.handleSectionToggle} settings={settings} />
         <ExtendedInfo accordionId="extendedInfoSection" stripes={stripes} user={user} expanded={this.state.sections.extendedInfoSection} onToggle={this.handleSectionToggle} />
         <ContactInfo accordionId="contactInfoSection" stripes={stripes} user={user} addresses={addresses} addressTypes={this.addressTypes} expanded={this.state.sections.contactInfoSection} onToggle={this.handleSectionToggle} />
         <ProxyPermissions accordionId="proxySection" onToggle={this.handleSectionToggle} proxies={proxies} sponsors={sponsors} expanded={this.state.sections.proxySection} {...this.props} />